-compartmentalized each class of components/__init__.py and taskchampion.py within the components folder. -added nescessary import statements for each new class file created
